samples/rag-demo-faiss/requirements.txt
faiss-cpu==1.7.4
openai==1.35.15
flask==2.3.3
numpy==1.24.3